python/src/requirements.txt
deepgram-sdk
requests

numpy
numba
scipy
librosa==0.9.2
torch
unidecode

wave
